вроде полностью рабочий интерфейс

+++ Assets/Scripts/Resources/Models/Resource/IResourceData.cs
@@ -9,7 +9,7 @@ namespace Resources.Models.Resource
         ICityData ConnectedCity { get; set; }
         void ConnectCity(ICityData city);
         void DisconnectCity();
-        int Quantity { get; set; }
+        float Quantity { get; set; }
         float Level { get; set; }
         int IntLevel { get; }
         void Harvest();
